fix resetGroupNodes except for sealer/observer
